requirements.txt
fastapi==0.114.0
langchain==0.2.16
langchain_community==0.2.16
langchain_core==0.2.38
openai==1.44.0
pydantic==2.9.0
python-dotenv==1.0.1
uvicorn==0.30.6
